Add vscode setting for phabricator URL fallback + differentiate fork
Added `open-in-diffusion.defaults.phabricator-uri` setting as fallback
for when `.arcconfig` doesn't contain `phabricator.uri`.

Updated the readme.md with this setting + differentiated this fork from
the original source.
Changed the extension author so it's more visible in the UI when looking
at installed extensions. Don't plan on publishing this anyway so it
shouldn't be visible to users other than me building it manually and
installing it from file.