Add option workspace-root-is-repo-root
If true, use the commented out code that assumes the root of the
workspace is the root of the repo. This makes getting the `.arcconfig`
file much much faster for bit repos.
If false, use the old glob search for `.arcconfig` file. This can detect
the repo root even if in a subfolder of the workspace, but is slow if
workspace has a lot of folders.